Implement favorites_add
