+++ DuckDuckGo/NewTabPage/Favorites/NewTabPageFavoritesClient.swift
@@ -33,7 +33,6 @@ final class NewTabPageFavoritesClient: NewTabPageScriptClient {
         self.faviconManager = faviconManager
 
         favoritesModel.$favorites.dropFirst()
-            .receive(on: DispatchQueue.main)
             .sink { [weak self] favorites in
                 Task { @MainActor in
                     self?.notifyDataUpdated(favorites)
@@ -42,7 +41,6 @@ final class NewTabPageFavoritesClient: NewTabPageScriptClient {
             .store(in: &cancellables)
 
         favoritesModel.$showAllFavorites.dropFirst()
-            .receive(on: DispatchQueue.main)
             .sink { [weak self] showAllFavorites in
                 Task { @MainActor in
                     self?.notifyConfigUpdated(showAllFavorites)
@@ -52,6 +50,7 @@ final class NewTabPageFavoritesClient: NewTabPageScriptClient {
     }
 
     enum MessageName: String, CaseIterable {
+        case add = "favorites_add"
         case getConfig = "favorites_getConfig"
         case getData = "favorites_getData"
         case move = "favorites_move"
@@ -64,6 +63,7 @@ final class NewTabPageFavoritesClient: NewTabPageScriptClient {
 
     func registerMessageHandlers(for userScript: any SubfeatureWithExternalMessageHandling) {
         userScript.registerMessageHandlers([
+            MessageName.add.rawValue: { [weak self] in try await self?.add(params: $0, original: $1) },
             MessageName.getConfig.rawValue: { [weak self] in try await self?.getConfig(params: $0, original: $1) },
             MessageName.getData.rawValue: { [weak self] in try await self?.getData(params: $0, original: $1) },
             MessageName.move.rawValue: { [weak self] in try await self?.move(params: $0, original: $1) },
@@ -71,6 +71,11 @@ final class NewTabPageFavoritesClient: NewTabPageScriptClient {
             MessageName.openContextMenu.rawValue: { [weak self] in try await self?.openContextMenu(params: $0, original: $1) },
             MessageName.setConfig.rawValue: { [weak self] in try await self?.setConfig(params: $0, original: $1) }
         ])
+    }
+
+    func add(params: Any, original: WKScriptMessage) async throws -> Encodable? {
+        await favoritesModel.addNew()
+        return nil
     }
 
     func getConfig(params: Any, original: WKScriptMessage) async throws -> Encodable? {
